Cleaning up code and documenting functions.
